case_study/04_youtube_playlist_qna_bot/requirements.txt
chromadb
youtube_transcript_api
llama-index-llms-gemini
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-readers-youtube-transcript
llama-index-core
google-api-python-client
speechrecognition
pyaudio
gradio
openai-whisper
